collapse open and close handlers
