+++ packages/cli/src/commands/wallet/lend/close.ts
@@ -1,53 +1,11 @@
-import { walletContext } from '@/context/walletContext.js'
-import { rethrowAsCliError } from '@/output/errors.js'
-import { printOutput } from '@/output/printOutput.js'
-import { collectMarkets, resolveMarket } from '@/resolvers/markets.js'
-import { parseAmount } from '@/utils/parseAmount.js'
-import { ensureOnchainSuccess, toReceiptArray } from '@/utils/receipts.js'
+import { type LendActionFlags, runLendAction } from './runLendAction.js'
 
-import { requireLendCapability } from './requireLendCapability.js'
-
-export interface LendCloseFlags {
-  market: string
-  amount: string
-}
+export type LendCloseFlags = LendActionFlags
 
 /**
- * @description Handler for `actions wallet lend close --market <name>
- * --amount <n>`. Resolves the market through the config allowlist,
- * delegates to `wallet.lend.closePosition`, and emits a structured
- * receipt envelope. The amount is the human-readable quantity to
- * withdraw - the SDK converts to wei. Reverts surface as `onchain`;
- * RPC failures as retryable `network`.
+ * @description Handler for `actions wallet lend close --market <name> --amount <n>`. Delegates to `runLendAction('close', flags)` which resolves the market through the config allowlist, dispatches `wallet.lend.closePosition`, and emits a `LendActionDoc` envelope. Reverts surface as `onchain`; SDK validation errors as `validation`; unknown failures as retryable `network`.
  * @param flags - Commander-parsed required options.
- * @returns Promise that resolves once stdout has been written.
  */
 export async function runWalletLendClose(flags: LendCloseFlags): Promise<void> {
-  const { wallet, config } = await walletContext()
-  requireLendCapability(wallet)
-  const market = resolveMarket(flags.market, collectMarkets(config))
-  const amount = parseAmount(flags.amount)
-  try {
-    const receipt = await wallet.lend.closePosition({
-      asset: market.asset,
-      marketId: { address: market.address, chainId: market.chainId },
-      amount,
-    })
-    const receipts = toReceiptArray(receipt)
-    ensureOnchainSuccess(receipts)
-    printOutput('lendClose', {
-      action: 'close',
-      market: {
-        name: market.name,
-        address: market.address,
-        chainId: market.chainId,
-        provider: market.lendProvider,
-      },
-      asset: { symbol: market.asset.metadata.symbol },
-      amount,
-      transactions: receipts,
-    })
-  } catch (err) {
-    rethrowAsCliError(err)
-  }
+  await runLendAction('close', flags)
 }
