+++ packages/cli/src/commands/wallet/lend/open.ts
@@ -1,54 +1,11 @@
-import { walletContext } from '@/context/walletContext.js'
-import { rethrowAsCliError } from '@/output/errors.js'
-import { printOutput } from '@/output/printOutput.js'
-import { collectMarkets, resolveMarket } from '@/resolvers/markets.js'
-import { parseAmount } from '@/utils/parseAmount.js'
-import { ensureOnchainSuccess, toReceiptArray } from '@/utils/receipts.js'
+import { type LendActionFlags, runLendAction } from './runLendAction.js'
 
-import { requireLendCapability } from './requireLendCapability.js'
-
-export interface LendOpenFlags {
-  market: string
-  amount: string
-}
+export type LendOpenFlags = LendActionFlags
 
 /**
- * @description Handler for `actions wallet lend open --market <name>
- * --amount <n>`. Resolves the market through the config allowlist,
- * delegates to `wallet.lend.openPosition` (which dispatches an optional
- * ERC-20 approval + the position call as a single sendBatch on EOA), and
- * emits a structured receipt envelope. Reverts surface as `onchain`;
- * RPC failures as retryable `network`. Both flags are enforced as
- * required by commander, so the handler trusts they are present.
+ * @description Handler for `actions wallet lend open --market <name> --amount <n>`. Delegates to `runLendAction('open', flags)` which resolves the market through the config allowlist, dispatches `wallet.lend.openPosition`, and emits a `LendActionDoc` envelope. Reverts surface as `onchain`; SDK validation errors as `validation`; unknown failures as retryable `network`.
  * @param flags - Commander-parsed required options.
- * @returns Promise that resolves once stdout has been written.
  */
 export async function runWalletLendOpen(flags: LendOpenFlags): Promise<void> {
-  const { wallet, config } = await walletContext()
-  requireLendCapability(wallet)
-  const market = resolveMarket(flags.market, collectMarkets(config))
-  const amount = parseAmount(flags.amount)
-  try {
-    const receipt = await wallet.lend.openPosition({
-      asset: market.asset,
-      marketId: { address: market.address, chainId: market.chainId },
-      amount,
-    })
-    const receipts = toReceiptArray(receipt)
-    ensureOnchainSuccess(receipts)
-    printOutput('lendOpen', {
-      action: 'open',
-      market: {
-        name: market.name,
-        address: market.address,
-        chainId: market.chainId,
-        provider: market.lendProvider,
-      },
-      asset: { symbol: market.asset.metadata.symbol },
-      amount,
-      transactions: receipts,
-    })
-  } catch (err) {
-    rethrowAsCliError(err)
-  }
+  await runLendAction('open', flags)
 }
